requirements.txt
APScheduler==3.7.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cloudflare==2.8.15
cryptography==3.4.7
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
jsonlines==2.0.0
namesgenerator==0.3
protobuf==3.17.3
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.17.1
python-telegram-bot==13.6
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
yandexcloud==0.90.0
